# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /practice/sessions:
    post:
      operationId: startPracticeSession
      tags: [practice]
      summary: Start a new practice session (infinite, adaptive)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeSessionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
  /practice/sessions/{sessionId}/next:
    post:
      operationId: nextPracticeProblem
      tags: [practice]
      summary: Generate the next problem (adaptive difficulty, de-novo)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NextProblemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeProblem"
  /practice/sessions/{sessionId}/grade:
    post:
      operationId: gradePracticeAnswer
      tags: [practice]
      summary: Grade a practice answer; returns correct/incorrect and explanation
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeAnswerInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeGrade"
  /tutor/ask:
    post:
      operationId: askTutor
      tags: [tutor]
      summary: Ask the AI tutor a question (text reply with optional spoken audio URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorAskInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorReply"
...
components:
  schemas:
...
    KeystrokeTrace:
      type: object
      required: [keystrokeCount, eraseCount, durationMs]
      properties:
        keystrokeCount: { type: integer }
        eraseCount: { type: integer }
        bulkInsertCount: { type: integer, default: 0 }
        longestBulkInsertChars: { type: integer, default: 0 }
        rewriteSegments: { type: integer, default: 0 }
        durationMs: { type: integer }
...
    PracticeSessionInput:
      type: object
      required: [tutorEnabled]
      properties:
        weekNumber: { type: ["integer", "null"] }
        topicId: { type: ["integer", "null"] }
        tutorEnabled: { type: boolean }
        focusOnWeaknesses: { type: boolean, default: true }
        initialDifficulty:
          type: ["number", "null"]
          description: Starting difficulty 1-5; if omitted defaults to 2.0
    PracticeSession:
      type: object
      required: [id, tutorEnabled, difficulty, weekNumber]
      properties:
        id: { type: integer }
        tutorEnabled: { type: boolean }
        difficulty: { type: number }
        weekNumber: { type: ["integer", "null"] }
        topicId: { type: ["integer", "null"] }
        focusOnWeaknesses: { type: boolean }
    NextProblemInput:
      type: object
      properties:
        request:
          type: ["string", "null"]
          description: Free-form user request, e.g. "give me harder probability problems"
        topicId: { type: ["integer", "null"] }
    PracticeProblem:
      type: object
      required: [id, prompt, topicId, difficulty]
      properties:
        id: { type: integer }
        prompt: { type: string }
        topicId: { type: integer }
        topicTitle: { type: string }
        difficulty: { type: number }
    PracticeAnswerInput:
      type: object
      required: [problemId, answer, trace]
      properties:
        problemId: { type: integer }
        answer: { type: string }
        trace:
          $ref: "#/components/schemas/KeystrokeTrace"
    PracticeGrade:
      type: object
      required: [problemId, correct, explanation, newDifficulty]
      properties:
        problemId: { type: integer }
        correct: { type: boolean }
        correctAnswer: { type: string }
        explanation: { type: string }
        newDifficulty: { type: number }
        tutorTip: { type: ["string", "null"] }
    TutorAskInput:
      type: object
      required: [message]
      properties:
        sessionId: { type: ["integer", "null"] }
        message: { type: string }
        selectedLectureText: { type: ["string", "null"] }
        lectureId: { type: ["integer", "null"] }
        problemId: { type: ["integer", "null"] }
    TutorReply:
      type: object
      required: [text]
      properties:
        text: { type: string }
        audioUrl: { type: ["string", "null"] }
...